for #2/1.2

+++ trap.c
@@ -86,11 +86,17 @@ trap(struct trapframe *tf)
               tf->trapno, cpu->id, tf->eip, rcr2());
       panic("trap");
     }
+	
     // In user space, assume process misbehaved.
-    cprintf("pid %d %s: trap %d err %d on cpu %d "
-            "eip 0x%x addr 0x%x--kill proc\n",
-            proc->pid, proc->name, tf->trapno, tf->err, cpu->id, tf->eip, 
-            rcr2());
+	// Do not echo exception when it is a problem of "return 0;" by user procedure.
+	// The second judgement is for whether it is a ret instruction or a jmp instruction
+	if((tf->eip+1)!=0 || (*(int *)(tf->esp-4)+1)!=0)
+		cprintf("pid %d %s: trap %d err %d on cpu %d "
+			    "eip 0x%x addr 0x%x--kill proc\n",
+				proc->pid, proc->name, tf->trapno, tf->err, cpu->id, tf->eip, 
+				rcr2());
+	
+		
     proc->killed = 1;
   }
